Implementing a much cleaner handling of auxiliary variables by exploiting the fact that we don't re-use pointers in the course of a simulation. The memory is allocated and the pointer is fixed, meaning an equation object can set these pointers at t = 0 and use them for all time.
In order for this approach to work in general, another restructuring of PKPM was needed, as PKPM has a very tight coupling between the kinetic species and the fluid species, with the time-stepping loop bouncing back and forth between the two structs to compute various quantities which were stored in each of the structs. Now, to facilitate this overall cleaner handling of the auxiliary variables, all PKPM variables are stored in the kinetic species struct and the fluid species struct is used only to update the fluid variables.

This approach relies on the currently implemented program flow that the field object is allocated first, then the species objects, then the fluid species objects, so that the fluid species can also pre-set the auxiliary variables pointers by just pointing to the kinetic species structs stored pointers (because these have already been allocated because the kinetic species is allocated before the fluid species).

As part of this restructuring, I have attempted to make the program flow as readable as possible by inserting the PKPM variable computations where they are needed, called only via handling of the kinetic species struct, with the handling of diagnostics like integrated variables and I/O now also exclusively handled by the kinetic species. While we will continue to exploit the common infrastructure of vm_fluid_species for the PKPM fluid object, the program flow is meant to emphasize that this is really a kinetic model, with the kinetic model computing almost everything and handling the I/O and diagnostics of the fluid variables.