Making polynomial order 2 OT problem look like polynomial order 1. There's a slightly subtlety right now that requires some testing, that to avoid extra communication/BC calls, we perform certain PKPM operations over the extended range, which includes corners. While these corners are not used, they can't be 0, because otherwise batched linear solves fail due to one of the matrices being singular.
It is likely that, in general, the parallelism afforded by just doing a straight accumulate and scale of the forward Euler update means that we don't have to do these over the local range. But we need to check this. For now, moving to a cluster to see if p=2 PKPM simulations are even stable