Some recent testing have revealed that div(b) when computed with recovery has significant oscillations in some PKPM tests. On one hand this is to be expected, and we had moved away from recovery on other terms to great robustness benefit.
However, for awhile we have resisted removing recovery from the computation of b_hat and div(b) because it was not obvious what we should do when these terms show up so similar to classical streaming (which can be handled basically trivially since v is a coordinate). I have elected to try something, which preliminarily works fine and may be a significant improvement, based on the success in changing u.

We compute b_hat at the interface and then utilize a modified Lax flux which reduces to streaming in the limit that b_l = b_r (b on the left and right of the interface). Further, it has the same split form structure that allowed us to achieve phase space incompressibility in the weak sense for u.

Right now, I just compute b and bb normally, and then evaluate them at the interfaces. For completeness, I should probably do what we did for u (evaluate rhou and rho at the interface, compute u) so that all the desired properties of b at the interface are maintained. For now, commit as is so we can start higher dimensional testing to see what difference this makes for the hardest tests we have of the PKPM system.