The transition to the new form of the fluxes involving b_hat permits a new, more accurate computation of b_hat at the surfaces. Now we compute BB at the surface, compute |B|^2, compute bb, and then compute the component of b needed on the interface (bx for x, by for y, bz for z).
In so doing, even in regions of the domain which previously needed to zero out all the high order components of bb and b for stability (so that we don't take the square root of a negative number and generate an imaginary b), we are guaranteed at least linear variation across a cell of the needed quantities. This linear variation is guaranteed because the cell average of Bx^2, By^2, and Bz^2 is guaranteed to be positive at the surfaces, and will be different at the left and right interfaces, thus granting us an implicit linear variation.