Now PKPM uniformly uses derived quantities computed at surfaces for all surface integrals in the DG algorithm. The final change was to add to the pressure tensor computation a "p_ij_surf" computation which computes the needed pressure tensor components at the corresponding surfaces (Pxx, Pxy, and Pxz at the left and right x interfaces for example). This procedure utilizes the previously implemented b_i b_j computed at the surface (which we were using anyway in computing b_hat at surfaces to increase the accuracy of the b_hat computation).
Modulo final debugging checks (valgrind clean and some GPU testing) PKPM in its current state (only the 0th Fourier harmonic and 0th and 1st Laguerre) is feature complete. There is an additional experiment which could be done based on what entropy-stable folks do, where we would change volume terms to use Gauss-Lobatto quadrature of the necessary order to include the computed surface expansions in the volume update, but such a change is non-trivial and at first glance is not guaranteed to buy us anything (it may in fact make things worse as the properties we worked to insure volume-wise and surface-wise are not point-wise properties, i.e., while we have ensured that rhou = rho * u and b . b = 1 within the volume and at surfaces, these quantities are not true point wise).

I may also get rid of the positivity check in the primitive variables for now. I realize if we reduce the order of the primitive variable solve, we introduce spurious flows into the problem which at first glance are just as bad as negative density. This is because we need the whole expansion of 1/rho div(p_par b) and T_perp/m to ensure the cancellation of terms to keep the first moment of the kinetic equation identically zero. I will examine further prior to the PR