During the course of the bhat-fix development, there was a point (commit 8cbf3d9) after the discovery of the missing mass factor in the consistent div(p_par b) pressure force in which low-guide field force-free reconnection stably ran well past the point of peak reconnection.
Unfortunately, in the subsequent clean up of pkpm over the next month in preparation for the bhat-fix branch merge, low guide field reconnection became less robust (still much more robust than before all the bhat-fix work, but less robust than commit 8cbf3d9).

There were two further core changes in the subsequent clean up. The first was the final elimination of recovery from the PKPM algorithm so that the parallel streaming is handled with a modified Lax flux: vpar/2 ( {{b}}{{F}} - max(|b_i|)/2 sign(vpar) (F^+ - F^-) ), which reduces to our upwind flux in the limit that b_l = b_r. The second was a switch to computing *surface* expansions of derived quantities such as b, u, p_ij, etc, under the logic that because all these derived quantities are computed with weak equality (especially weak division in many places), a more accurate representation of these quantities is to evaluate the base quantity at the relevant surfaces before doing the weak operations.

This branch will explore reverting the second development. It is possible the second development is less robust without corresponding changes to the volume kernels, since the volume kernels are inclusive of the boundaries but in the current formulation with the "more accurate" surface expansions, the interior volumes are, in a sense, disconnected from the edges. In fact, this disconnection may be why all the robust entropy-stable DG schemes use high order Gauss-Lobatto quadrature, so that quantities computed at the surface are included in the volume integral evaluation. More testing is needed (moving to Frontera and GPU machines to run force-free).

Right now just changing b_hat. If this improves things, we will change u and p_ij as well to match this program flow