Complete rework of surface expansions for robustness testing. Now, after the weak operations are done on the volume expansions of bb, b, u, and p_ij, we immediately evaluate these quantities at the needed surfaces (keeping the same program flow from before when we did linear solves for these variables).
The most important robustness improvement may be the elimination of the use of the cell_avg flag from the primitive variables. Preliminary testing in main suggests actually the second this flag reduced the order of the primitive variables, the simulations would quickly crash. Intuitively, this would be because we rely on 1/rho * rho = 1 so that now spurious flows develop due to the pressure forces, and if we throw away the high order information in 1/rho, this will lead to an error propagating through the algorithm. Of course if rho becomes negative that is unphysical and can crash a simulation in its own right, but it is interesting that is better to take a timestep with some amount of negative rho rather than to reduce the order of the weak division.

In general, a true positivity fix is needed.