Through further testing, I discovered something subtle. Originally, I believed the LBO in the PKPM model had a simpler structure than the typical Vlasov LBO owing to the fact that the PKPM model is in the local flow frame.
Thus, one would think div((v - u) f) -> div(v'f) for the drag term.

This statement is only true in the continuous limit though, and in fact the use of our conservative LBO algorithm (including the boundary corrections) was creating a spurious flow velocity in all simulations. In many cases, this spurious flow velocity was small (for example, the classical neutral sod shock was generating a 1.0e-7 flow velocity from the collision operator). But for other cases, such as shock-waves, the collision operator would excite a flow which would then crash the simulation (only discovered because I, on a whim, turned off collisions in one shock simulation and found it to be more robust).

So, the PKPM LBO has been changed to a more familiar form, similar to what is employed in Vlasov, with a primitive moment solve for a delta u correction that then allows us to keep the M1 = integral(vpar F_0) close to machine precision.