Via some testing this weekend, I realized that the gradient-based closure was not actually parallelized, defaulting to always looping over the global range ro compute the heat flux, even if the local range for the app was still utilized to construct the final update to the pressure.
In the process I realized this whole time the gradient-based closure had a significant error: the non-ideal variable range wasn't actually being used for updating the pressure tensor, so the div(q) was wrong in the final right-most cell. This error of course affects simulations performed, though perhaps only quantitatively and not qualitatively (the parallel firehose run still grows a mode with basically the same wavelength). More testing is required.

For right now, I use the create_ranges method from rect_decomp which is used based on the input range from the user-side. What this does is create a local range for the non-ideal variables with one extra "cell" on each side so that the non-ideal variables still correspond to vertex values, we are just doing one extra operation than needed (we only need one extra "cell" not one extra cell on each side; since the ranges are employed consistently we do in fact still get the same answer, we just do a tiny bit more work).