Experimenting with a G0 app refactor based on issues we've been discussing about how different certain apps are. The basic idea is to separate the PKPM model into its own app in its entirety, owing to the additional complexity of solving a momentum equation paired with a kinetic equation, and the fact that these two objects are not separate but instead a joint unit. Not tested yet but does compile.
In the process I decided that if PKPM is going to be entirely separate, to try to streamline its naming conventions. Need to be careful that nothing has changed computation-wise and this is just a refactor on the app side (with the needed changes in the zero directory for the updaters).