Utilizing the infrastructure change to hyper DG for GK which computes (at least part of) the stable time step from the surface kernels, compute certain components of the stable timestep for PKPM from the surface kernels.
Since two pieces of the surface kernels involve the velocity coordinate, I now use the surface kernels to get the actual corner value of the velocity (and thus weight the velocity-dependent part of the advection with vmax instead of vmax - dv/2) along with an estimate of the variation of certain forces (such as the pressure force) in the cell from the needed quadrature point evaluations.

Time step estimate may be slightly aggressive because it uses vmax b_i + u_i + sqrt(3 P_ii/m) by reusing the Lax penalization, but for many simulations this won't matter because the timestep is set by the speed of light. Where this is really necessary is during short periods where some transient phenomena makes the kinetic equation set the timestep (due to large variation of vpar bb : grad(u)) and now we can run with cfl_frac = 1.0 and take a smaller timestep during the transient period, but relax to a larger timestep after the transient dies away.