In the input file revamp for the PKPM app, I assumed all of the robustness improvements to PKPM would allow for a more traditional shock input file with a jump in the initial flow velocity. However, if I use this traditional initial condition du/dx at t = 0 is huge and immediately spikes particles into the zero flux boundary.
So I have switched back to a tanh(x) profile initial condition. It is a very sharp tanh (on the scale of the Debye length) so the shock looks basically identical to a traditional Vlasov initial condition, but it is much more stable because du/dx at t = 0 is well behaved. Almost certainly further indicates the need for limiters.