Fixing GPU build. This fix required some rearrangement of where certain functions are defined for the pointer handling, specifically the acquire methods, which we now want to use on the device to properly handle the memory (so a launched CUDA kernel acquires the pointer to a particular object, such as an array, wv_eqn, or wave_geom object). Also fixing calls to GKYL_MAX2 and GKYL_MIN2.
I can see no immediate disadvantage to the way that I fixed things, moving certain acquire methods from the .c file to the .h file, but perhaps some further discussion should be had on whether this is the correct rearrangement to make the object visible on the device side. Will do further testing.