Experimenting with a proposed reorganization of the gyrokinetic collisionless kernels for performance. Originally we consolidated all of GK (including its myriad of different options: slab, helical, general geometry, EM, etc) into a single set of kernels which would work across all cases.
Some experimentation has revealed this has some significant inefficiencies though for problems of interest. The problem is that the computation of alpha, the phase space flux, as alpha = dz/dt = {z, H} is a non-trivial computation. While the composition of alpha (which coefficients of alpha's expansion are non-zero) is relatively agnostic to all these different options (for example, alpha will always have a number of non-zero coefficients due to grad(phi) acceleration in vparallel and grad(phi) x b E x B drift in configuration space), the number of computations we do for any given GK configuration is very different for different GK models and the computation of alpha is a significant percentage of the overall update (upwards of 40% of the floating point operations performed). Further, we are computing alpha more times than we need to in a cell-based update because alpha is continuous (or mostly continuous, Aparallel introduces some complications we can think about through this refactor).

So, I have re-done all the GK collisionless kernels to do the following:
1. At t=0, I compute the time-independent piece of Bstar/Bmag (the part not dependent on Aparallel)
2. At every time step, I loop over the extended phase space domain and compute the surface expansion of alpha via the Poisson bracket, storing the surface expansion on the *lower* edge following the convention in the Gkeyll fluids solver for the storage of surface quantities (and utilizing the extended phase space range to obtain the surface expansion of alpha everywhere needed)
3. Pass this surface alpha expansion to the surface kernels

By storing the surface alpha expansion, there is no large increase in the amount of memory (and also the surface expansion can be the same accuracy as what we were currently doing, evaluating {z,H} and obtain a high order polynomial, then projecting that high order polynomial onto surface bases). Further, I utilize the reduced variable space of Bstar/Bmag (the time-independent piece at most depends on (x,z,vpar)) to make an efficient storage and re-use of this fairly dense object, which should also especially help on GPUs by reducing register pressure.

Further thoughts:
1. We should consider having a set of kernels for by = 0 (no toroidal field) for use case in LAPD or mirror simulations, as this will further reduce the number of computations needed to find alpha, even if the overall sparsity of alpha is unchanged because it always has non-zero terms in (x,y,z,vpar,mu) from the various pieces of the Hamiltonian.
2. The separation of alpha out like this also allows for a design of the GPU kernels that further utilizes the GPU, where we can spawn threads equal to 4*number of phase space cells to saturate the GPU more fully on coarse grids.
3. The inclusion of Aparallel can be handled directly within the kernels or within the alpha computation, allowing for a reasonable separation between ES and EM runs without a combinatorial explosion of kernels.