Experimenting with another iteration of the gyrokinetic surface fluxes which will hopefully further improve robustness and reduce costs.
For a few years now, we have defined a modal upwind flux where we evaluate the surface phase space flux at quadrature points, construct an "upwinded distribution function" in our modal expansion by choosing f^+ or f^- at quadrature points and transforming between nodal and modal bases with the proper choice of nodal basis (nodal basis at Gauss-Legendre quadrature points, which has an alias-free transformation between the nodal and modal bases). This flux is alias-free within the approximation of the distribution function at the interface as this "fUpwind." It was a *significant* improvement over the penalization fluxes we utilized at the beginning of G2 and the gyrokinetic updaters success with these fluxes propagated into their use for other kinetic equations.

But there is another formulation of a similar idea which is, at first glance from the L2 stability proof, *more robust*. We instead formulate the numerical flux as:
F = alpha_surf*(1/2 (f^+ + f^-) - sign(alpha_surf)/2 *(f^+ - f^-))

where the same lessons and experience for transforming between nodal and modal bases are used to construct sign(alpha_surf) as a modal expansion.

We also save on cost by defining a boolean variable which checks if sign(alpha_surf) is single-signed (either +1 or -1) and then just using a traditional upwind flux.

More testing needs to be done, and the Maxima to generate these kernels needs cleanup, but this could be a potentially powerful way of formulating our surface fluxes for any general system in which the flux is continuous (such as all our kinetic equations).