This commit has two key features, one of which is a potential robustness fix of the GK kernels which could have been causing issues for some GK simulations, and the other is a feature extension.
The robustness fix is a fix to a potentially serious issue based on how we compute alpha that I only discovered because of our refactor of the computation of alpha: our use of the intermediate variable for B*/Bmag. The use of this intermediate variable (which has been in the G0 kernels for awhile) has the potential to make alpha discontinuous at cell interfaces. The reason is the same reason we compute the surface alpha the way we do.

alpha = dz/dt = {z, H}, where { } is the Poisson bracket

From, e.g., Noah's thesis, we can see that the computation of the Poisson bracket produces a much higher order polynomial than we represent with our basis expansion. Thus, in order to guarantee alpha is continuous, we compute the Poisson bracket with its very high order expansion and then evaluate this expansion at the needed surface before projecting onto our surface basis. So for example, there is a piece of the Poisson bracket that goes like B*/Bmag . (grad(F) dG/dvpar - grad(G) dF/dvpar), which could produce up to a cubic polynomial that is then evaluated at the surface.

EXCEPT, B*/Bmag is itself a nonlinear product because of factors like the inverse Jacobian! So our downprojection of B*/Bmag does not guarantee continuity. We must instead leave B*/Bmag as its total nonlinear expansion as part of the construction of alpha at the surface.

Thus, I am deleting some of the work I did to pre-compute B*/Bmag along with a refactor of the kernels to never compute the expansion of B*/Bmag and just leave it as its nonlinear product. It is possible we could still use an intermediate variable for the volume expansion but for now, I don't utilize an intermediate variable for this anymore, not when I know this was wrong before.

The fact that we can't pre-compute this factor does mean certain kernels such as 3x2v p=2 Serendipity are prohibitively large, and since right now no one is running 3x2v p=2 in production, I have removed these kernels from the gyrokinetic kernels directory. This also speeds up compilation and I should follow through and do this consistently across GK (no LBO 3x2v p=2 kernels either).

The feature extension is the addition of a set of kernels for the no toroidal field limit (by=0) applicable to mirrors, which will help bring down the cost now that we have to do something slightly more expensive not computing the intermediate B*/Bmag variable. These no toroidal field kernels also have the benefit of making slab calculations like LAPD much cheaper too. There are some further performance improvements that could be made to the no toroidal field kernels (the surface fluxes in x and z are *super* sparse but their sparsity is not being completely exploited right now because it requires a careful look at the Maxima for how to handle utilizing a basis which is as sparse as it can be without introducing needless complication in the size/structure of the arrays). But I will delay these until I have a better idea of how much cheaper the no toroidal field kernels are.