Proposed re-doing of the 2x2v kernels to accommodate a desired use case: axisymmetric calculations.
Right now in main, 2x2v kernels are an esoteric use case of the actual field (that one would typically be following and that you have parallel streaming to) is out of the plane, with Bstar/Bmag only having their x and y components (things like the curvature drift). As far as I am aware, there are no production use cases for this that cannot just use the Vlasov Poisson kernels with magnetic field out of the plane.

In this proposed rearrangement, 2x2v becomes a set of kernels solving the GK equation in an (x,z,vpar,mu) set of coordinates, where z if the field line coordinate and x is the radial coordinate.

To fully hook-in, we will need a new field solve which solves a 1D Poisson solve for the perpendicular (x) dynamics, but this should have the advantage of being much more efficient than the 2D Poisson solves.

Of further utility: the no toroidal field, by=0, kernels in this use case serve two purposes. The first is that for a mirror, one can now do an equilibrium calculation in the GK limit without having any free parameters on the Poisson solve. The second is that, coincidentally, for by=0 there are no drifts and thus we can achieve a desired functionality for our collaborators to create a "kinetic SOLPS" that does an axisymmetric calculation without any drifts.