Refactoring the GK radiation operator to follow some Gkeyll conventions and make the process of debugging a bit easier by allowing the utilization of the radiation operator through the G0 app. The biggest change is the movement of the computation of the two drag coefficients out of the equation object into a "dg_calc_gk_rad_vars" updater which follows the conventions employed elsewhere for constructing continuous functions using eval_on_nodes.
These coefficients are computed at t=0 and then stored. Some further work is needed because we compute the radiation operator for the electrons for every ion species present (since this operator is approximating things like bremmstrahlung radiation and is multiplied by the ion density). But we can follow cross collisions infrastructure and construct a total radiation operator for all the coefficients present (so any number of ion species).

Will also add a write method that allows us to write out vnu and vsqnu so we can see if these functions look like what we expect.