Working with Dingyun to construct a gyrokinetic BGK collision operator in the G0 app. Follows a similar structure to LBO, initializing the moments it needs and computing them depending on the details of the collisions (do cross-collisions exist, etc.).
Right now somewhat brute force, as it's allocating some additional distribution function arrays to handle the general case of an arbitrary number of cross collisions and doing the update of the RHS with simple accumulate calls. We may want to reformulate this to not allocate so much memory but I don't expect that to be an issue right now. Onwards to testing!