Adding adiabatic electron model and some other cleanup of the app. Lua wrapper now creates geometry input struct instead of expecting mapc2p and bmag functions. Also added the diffusion object.
An important subtlety I need to understand is how the ambipolar potential solve utilizes sheath values. There seems to be an assumption about how sheath values is organized as a data structure that I am not following, so I would not expect right now this object to work. Need to more carefully understand the expect behavior of ambi_bolt_potential.c