Refactor of general geometry neutrals based on discoveries with gyrokinetics for the best way to handle the surface alpha and new upwind fluxes.
Now as part of the initialization, we loop over the grid and get a surface alpha, the sign of surface alpha, a boolean for whether the surface alpha is single-signed, and a volume expansion of the cotangent vectors e^i.

The surface quantities are used in the surface update so that the general geometry neutrals updater looks structurally identical to the new gyrokinetic kernels. The volume expansion of the cotangent vectors e^i simplifies the volume kernels in general geometry neutrals by separating the integration of the volume term over velocity space and configuration space.

The signatures of streaming kernels for Vlasov have changed to accommodate general geometry neutrals, but no other changes were made to Vlasov kernels except for fSkin -> fskin and fEdge -> fedge in the boundary surface kernels.

I deleted the unit test for alpha_gen_geo and the files associated with that updater. More testing is needed of this new infrastructure but I am confident this is the direction we want to go