Committing the nearly complete skeleton of reactions based on how we want to refactor the reactions running through the G0 app. The steps are as follows: 1. Initialize every species and neutral species 2. Using the "cross_init" methods, sort the reactions based on the input struct for things like who is the electron species, who is the ion species, what kind of reaction are we doing, etc. 3. Within the forward Euler method, first call a "cross_moms" method similar to how we handle cross collisions with LBO. This method computes all the configuration space fields needed for each reaction present, including the reaction rate and other primitive moments that may be needed. 4. Within gk_species_rhs, call the react_rhs method which will do the needed update based on the enum "type_self" that says what the species in a given reaction (so all the phase space computations are handled within the app, computing things like the needed Maxwellians and the conf_phase weak multiplications).
Some more work is needed as we need to refactor dg_iz and dg_recomb to only compute moments, not the total phase space update. Further, we have moved enums from specific reaction objects to gkyl_eqn_type and this needs to be refactored in the init methods. But this is on the right track

I also did some simplifcation of the neutral species objects. While we need different methods for things like moments, sources, projection, and reaction, we do *not* need different neutral-specific structs besides the general gk_neut_species struct. This simplifies the private header slightly.