Refactoring the way we are doing radiation to reduce errors and the risk of radiation pushing particles the wrong direction (an incorrect characteristic in velocity space).
Now what we do is pre-compute two different expansions of the radiation drag. One in the volume (utilizing the more accurate Gauss-Legendre quadrature which does not guarantee continuity but is *only* used by the volume kernels) and one on vparallel and mu surfaces, because we can guarantee continuity with this pre-projection of the static function in velocity space.

As part of this refactor I did two additional improvements to the way we compute the radiation drag for quality of life and longevity: I utilize only B = B(x,z) when doing 3D simulations to reduce the memory cost in 3D when computing all the drag coefficients which might be present in a general simulation with lots of different sources of radiation. And I do the configuration space-phase space weak multiplication by the ion density in a kernel so that we do not need additional memory for the accumulation of all the drag coefficients (though we still utilize one set of additional phase space arrays now because n = n(x,y,z) so the final sum over drag coefficients is a full phase space array).

Checked the coefficients by eye and they look like the functional form in Jonathan's notes so hopefully this improves things. Even if this is *not* the source of the error, it is the right thing to do and will improve our lives once we find other bugs present and try to run production calculations.