Fixing the unit tests in preparation for final merge. This fixes all the seg faults and these tests now pass again.
There are two more long-term subtlties to address in the design of some of these unit tests. The first is that it would be better to go through and do a proper unit test of moments by taking exact solutions and comparing our projection routines against them for convergence and the like. I had to change some coefficients being checked in the unit test because the way geometry deflation works, the deflation is along x/y and the test was previously putting in certain Bmag factors by hand. Ideally one would develop some very non-trivial test with non-trivial n, upar, Tpar, and Tperp and check the results after utilizing the proj_bimaxwellian_on_basis or something.

Speaking of non-trivial, this brings me to the second subtlety. The correct_maxwellian test was giving failure results for "trivial" solutions because the iterative scheme would change the slopes slightly for a quantity that was constant (i.e., once you start manipulating the function, it's not actually constant, it has some small slope and teeny-tiny fluctuation in it). Basically a floating point comparison issue. Somehow we need to be more careful about this. Now that correct_maxwellian is doing a highly non-trivial test (trig function density, uniform flow, gaussian temperature) it passes with flying colors.