Taking a look at the unit and regression tests for radiation (along with the discovery that I was computing integrated moments incorrectly by not including the inverse Jacobian factor before integrating).
The radiation tests behave like there is some relatively large factor missing in the overall loss. After going through the tests and cleaning a few things up, I was able to get the disagreement on the order of O(1000) and it does converge to a number. Further the regression test behaves reasonably for finite time before it cools so much the distribution function can't be realized on the grid.

My suspicion is we are missing some factors somehow or have messed up the fit in some way (i.e., it's not a DG problem, but a problem with what we're putting in). The overall loss is negative like it should be and it displays reasonable properties like convergence, but it's not giving what we expect, suggesting we've messed up on the "what are we telling the system to project." The regression test has the plasma cooling ridiculously fast (on the timescale of 1.0e-8 the plasma is already hard to represent on the grid) so it would make sense if there was just a large constant factor missing in our implementation somehow.