BUG FIX BABY: There were two errors that seemed to intermix in a highly non-trivial way. The way CUB works is that it does a reduction over blocks and then it is left to a final atomic operation to do the reduction over threads. Problem: CUDA atomics for min and max don't support doubles, so we found some code online with the CUDA function atomicCAS for converting between long long and double and mixed that with the reduction over threads.
For God only knows why, this operation was not robust and did not actually always reduce over *all* the threads.

So I found some other code online similar to what atomicAdd does (which does natively support doubles) that does not mix the thread reduction and double conversion. So now every thread in the block (threadIdx.x < BLOCKSIZE) calls this atomicMax which simply does the atomicCAS and conversion call (with the if statement giving us the max operation).

From the looks of what I've written, this max operation assumes the number you're reducing is positive definite. I don't know whether that breaks something else in the code (I don't think so but I'm saying this out loud here in the commit message for someone to chime in)