Fixed several bad code practices. Please avoid these:
- Do not forget the "pragma once" directive on top of each header file (gkyl_nodal_ops.h)
- Do not include static methods in headers if they do any allocations (gkyl_read_radiation.h)
- Do not redefine two enum names. This is legal in C but not in C++ (gkyl_gkgeom.h)
- Do not include priv headers in public header. (gkyl_nodal_ops.h)

One thing to keep in mind: G0 is a C code and can't be compiled with a
C++ compiler. However, it is possible that the *compiled* G0 C library
may be *used* from C++. This is allowed by design. Recall you can call
C from C++. However, to ensure we can do this we MUST NOT use any
C-specific features in our public-facing headers (C-specific code in
private headers is fine)