Commiting a new infrastructure for a more generic handling of computing the "Maxwellian moments" or the moments of the equivalent Maxwellian (n, V_drift, T/m) where n and T/m are the stationary-frame density and temperature/mass and V_drift is the velocity of that stationary frame.
Now both non-relativistic and relativistic simulations can utilize and write out these "Maxwellian moments," with the appropriate changes for relativity directly encoded.

Not all the simplifications are there yet but this infrastructure should allow for cleaning up the correct_maxwellian Vlasov object as well as the BGK collision operator. Further, it allows for a new diagnostic "MaxwellianMoments" to be written out, which is especially useful in Vlasov calculations for generating an alias-free representation of the V_drift and temperature.

I will work with Grant to extend this infrastructure and hopefully build a GK version of this as well for writing out things like u_parallel, T_parallel, and T_perp as diagnostics. The FPO work John is doing will also benefit from this object.