Now that we have a streamlined capability for initializing the "LTE" distribution function in Vlasov, whether it's relativistic or non-relativistic, and inspired by the gk_species_projection object in the gk-g0-app branch, make a vm_species_projection object which initializes the LTE distribution (Maxwellian or Maxwell-Juttner) and corrects the moments (by default always corrects the density and has the option to correct all the moments).
Required some tedious changes to all the Vlasov input files and many of these should switch to initializing from a Maxwellian instead of the function projection, but this at least gets us started. I noticed actually as a result of this that sheath_1v and vm_sheath_1x1v_p2 were identical tests. Not sure when that happened. I deleted sheath_1v.