Refactoring how the projection routine for the LTE distribution function for Vlasov is computed now that we've got some nice, clean interfaces for switching between relativistic and non-relativistic systems.
The basic idea is two-fold:
1. There is a single projection routine for "Vlasov LTE" distribution functions and internal to the projection routine through a flag called "is_relativistic" it will either project a Maxwellian or a Maxwell-Juttner distribution (utilizing a flag instead of model_id since different models may desire the non-relativistic or relativistic distribution). We thus avoid the significant code duplication of having all the quadrature infrastructure defined for both Maxwell-Juttner and Maxwellian projection.
2. The proj_vlasov_lte routine *always* corrects the density of the projected distribution function. This correction is required anyway because to avoid finite precision effects, we do an expansion of the Bessel functions in the Maxwell-Juttner distribution so that we can obtain our desired temperatures, but which gets the density very wrong and thus requires the re-scaling.

This refactor now eliminates a significant percentage of the app-level Vlasov checking for relativistic vs. non-relativistic in places such as vm_species_projection and vm_species_bgk. This refactor also means that the correct_lte updater no longer needs a "correct density" advance method because the only place this would be called is in the projection routine and we don't want to do that because the correction routine itself needs projection routines (so if the correction routines creates a projection routine, and the projection routine creates a correction routine.... its turtles all the way down). I further added a "density-only" moments advance method to reduce the number of operations required to correct the density and avoid a current limitation of the code that we can't compute the relativistic pressure for cdim > 1 and poly_order > 1 due to the lack of the basis_inv method for those inputs.

Unit tests and pass and SR Neutral BGK test looks beautiful. Next steps:
1. Get everything on GPUs and run the SR Neutral BGK test on GPUs
2. Consolidate the naming convention so that the moments, projection, and correction routines all have vlasov_lte to emphasize how they are connected.